pyproject.toml
[tool.poetry]
name = "framewise-secureline"
version = "0.1.3"
description = "A client library for content security detection"
authors = ["Sigireddy Balasai <sigireddybalasai@gmail.com>"]
readme = "README.md"
packages = [{include = "framewise_secureline"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-vcr = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests"
]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=secureline --cov-report=term-missing"